Extract shared levenshtein utility from 9 duplicated validators
The same Levenshtein distance implementation was copy-pasted across
9 validate.py files. Extract it to common_scaffold/validate/levenshtein.py
and update all validators to import from the shared module.

No behavioral changes — identical algorithm, just deduplicated.

Files updated: GITHUB_REPOS/query2,4, PANCANCER_ATLAS/query2,
PATENTS/query2,3, stockmarket/query2,3,4,5
