+++ query_GITHUB_REPOS/query2/validate.py
@@ -1,25 +1,5 @@
 import re
-
-def levenshtein(s1, s2):
-    """
-    Standard Levenshtein distance implementation
-    """
-    if len(s1) < len(s2):
-        return levenshtein(s2, s1)
-    if len(s2) == 0:
-        return len(s1)
-
-    previous = list(range(len(s2) + 1))
-    for i, c1 in enumerate(s1):
-        current = [i + 1]
-        for j, c2 in enumerate(s2):
-            insert = previous[j + 1] + 1
-            delete = current[j] + 1
-            substitute = previous[j] + (c1 != c2)
-            current.append(min(insert, delete, substitute))
-        previous = current
-
-    return previous[-1]
+from common_scaffold.validate.levenshtein import levenshtein
 
 def validate(llm_output: str):
     """
