+++ query_PANCANCER_ATLAS/query2/validate.py
@@ -1,25 +1,9 @@
 import re
+from common_scaffold.validate.levenshtein import levenshtein
 
 def normalize(text: str) -> str:
     # Lowercase and collapse multiple whitespaces
     return re.sub(r'\s+', ' ', text.lower().strip())
-
-def levenshtein(s1: str, s2: str) -> int:
-    if len(s1) < len(s2):
-        return levenshtein(s2, s1)
-    if len(s2) == 0:
-        return len(s1)
-
-    previous_row = list(range(len(s2) + 1))
-    for i, c1 in enumerate(s1):
-        current_row = [i + 1]
-        for j, c2 in enumerate(s2):
-            insertions = previous_row[j + 1] + 1
-            deletions = current_row[j] + 1
-            substitutions = previous_row[j] + (c1 != c2)
-            current_row.append(min(insertions, deletions, substitutions))
-        previous_row = current_row
-    return previous_row[-1]
 
 def fuzzy_match(name: str, text: str, max_distance: int = 3) -> bool:
     name_len = len(name)
